requirements.txt
mysqlclient>=2.0.3,<3.0
Django==4.0.1
djangorestframework>=3.13.0,<3.16.0
django-environ>=0.5.0,<1.0.0
django-model-utils==4.2.0
django-storages[azure]==1.14.2
pillow
whitenoise
django-cors-headers
gunicorn


